+++ api/test/common/string_util_test.cc
@@ -6,7 +6,6 @@
 #include <string>
 
 #include <opentelemetry/common/string_util.h>
-#include "opentelemetry/nostd/string_view.h"
 
 // ------------------------- StringUtil class tests ---------------------------------